requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.0.0
numpy>=1.24.0

# Optional: For enhanced functionality
# chromadb>=0.4.0  # For persistent vector database
# langchain>=0.1.0  # For advanced RAG patterns
